[posix-sockets] byteswap port numbers.
port numbers are stored in network byte order, so we need to swap
them to host order on little endian platforms.

We don't call ntohs and friends from the OS here as that isn't
as easy to bind since it is a macro on some platforms. (And the
actual code required is very easy...)

* posix-sockets/utils.dylan
  (ntohs): New function.

* posix-sockets/posix-sockets.lid: Build utils.dylan

* posix-sockets/socket-address-inet.dylan
  (socket-address-port): Call ntohs to convert to host order.

* posix-sockets/socket-address-inet6.dylan
  (socket-address-port): Call ntohs to convert to host order.

* posix-sockets/library.dylan
  (library posix-sockets): Use system.

* posix-sockets/module-posix-sockets.dylan
  (module posix-sockets): Use operating-system to get
   architecture-little-endian?.